Add Limit option
Limit the number of versions of a package per distribution,
architecture, component, and type. The limit must be a number. If the
number is positive, all old package version that exceed these limit will
be removed when a new package version is added. If the number is zero or
negative, all package version will be kept. By default only one package
version will be kept.